+++ js/dustpress.js
@@ -1,173 +1,138 @@
-window.DustPress = ( function( window, document, $ ) {
-	var dp = {};
-	
-	dp.defaults = {
-		"type"	  		   : "post",
-		"tidy"    		   : false,
-		"render"  		   : false,
-		"partial" 		   : null,
-		"upload"  		   : false,
-		"data"             : false,
-		"url"              : null,
-		"bypassMainQuery"  : true,
-		"success" 		   : function() {},
-		"error"   		   : function() {},
-		"uploadProgress"   : function() {},
-		"downloadProgress" : function() {},
-		"contentType"      : "application/json"
-	};
+import 'whatwg-fetch'; // fetch polyfill
+import 'core-js/fn/promise'; // Promise polyfill
+import Cookie from 'js-cookie';
 
-	dp.start    = function() {};
-	dp.complete = function() {};
+export default class DustPress {
 
-	// Create token for CSRF protection
-	if ( typeof window.crypto !== 'undefined' && typeof window.crypto.getRandomValues === 'function' ) { 
-		dp.token = '';
+    /**
+     * CSRF token
+     *
+     * @static
+     * @type {String}
+     */
+    static csrfToken = '';
 
-		for ( var i = 0; i < 4; i++ ) {
-			dp.token += window.crypto.getRandomValues( new Uint32Array(1) );
-		}
-	}
-	else {
-		dp.token = Math.random() + '' + Math.random();
-	}
+    /**
+     * Default dpjs_token cookie params
+     *
+     * @type {Object}
+     */
+    cookieParams = {
+        expires: 60 * 60 * 1000,
+        path: '/'
+    }
 
-	dp.ajax = function( path, params ) {
+    /**
+     * Default dp() params
+     *
+     * @type {Object}
+     */
+    params = {
+        url: window.location,
+        method: 'POST',
+        args: {},
+        credentials: 'include',
+        headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json'
+        },
+        bypassMainQuery: false,
+        partial: '',
+        render: false,
+        tidy: false
+    }
 
-		var post = $.extend( {}, dp.defaults, params );
+    /**
+     * Generate token for CSRF protection
+     *
+     * @return {String} Generated token.
+     */
+    generateToken() {
+        // Use the already generated token if we have one
+        if ( this.csrfToken ) {
+            return this.csrfToken;
+        }
 
-		// Create a new instance of the default object so that simultaneous	calls wouldn't clash.
-		var instance = {};
+        let token = '';
 
-		// Assign the variables
-		instance.type               = post.type;
-		instance.success 	        = post.success;
-		instance.error 	            = post.error;
-		instance.url                = post.url;
-		instance.uploadProgress     = post.uploadProgress;
-		instance.downloadProgress   = post.downloadProgress;
-		instance.get 		        = post.get ? params.get : '';
-		instance.path		        = path;
-		instance.args               = post.args;
-		instance.data               = post.data;
-		instance.partial            = post.partial;
-		instance.render             = post.render;
-		instance.tidy               = post.tidy;
-		instance.upload             = post.upload;
-		instance.contentType        = post.contentType;
-		instance.bypassMainQuery    = post.bypassMainQuery;
-		instance.params 	        = params;
-		instance.start              = dp.start;
-		instance.complete           = dp.complete;
+        if ( typeof window.crypto !== 'undefined' && typeof window.crypto.getRandomValues === 'function' ) {
+            for ( let i = 0; i < 4; i++ ) {
+                token += window.crypto.getRandomValues( new Uint32Array( 1 ) );
+            }
+        } else {
+            token = Math.random() + '' + Math.random();
+        }
 
-		if ( instance.get.length && ! instance.get.startsWith('?') ) {
-			instance.get = '?' + instance.get;
-		}
+        this.csrfToken = token;
+        Cookie.set( 'dpjs_token', token, this.cookieParams );
 
-		
+        return token;
+    }
 
-		var date = new Date();
-        date.setTime( date.getTime() + ( 24*60*60*1000 ) );
-	    
-	    // Set the cookie for the token
-	    document.cookie = 'dpjs_token=' + dp.token + '; expires=' + date.toGMTString() + '; path=/';
+    /**
+     * Call dustpress function
+     *
+     * @param  {String} path   Dustpress call path.
+     * @param  {Object} params Call args.
+     * @return {Promise}       Result promise.
+     */
+    send( path, params = {}) {
 
-		var options = {
-			url: ( instance.url || dustpressjs_endpoint || ( window.location + instance.get ) ),
-			method: instance.type,
-			contentType: instance.contentType,
-			data: {
-				dustpress_data: {
-					path            : path,
-					args            : instance.args,
-					render          : instance.render,
-					tidy            : instance.tidy,
-					partial         : instance.partial,
-					data            : instance.data,
-					token           : dp.token,
-					bypassMainQuery : instance.bypassMainQuery
-				}
-			}
-		};
+        // Get the CSRF token
+        const token = this.generateToken();
 
-		// Stringify data so it can be sent
-		options.data = JSON.stringify( options.data );
+        // Extract params from the combination of default and given params
+        const {
+            url,             // Url to send ajax call to
+            method,          // Ajax call method
+            args,            // Dustpress ajax args
+            bypassMainQuery, // Whether the main wp query should be disabled
+            partial,         // What partial to render
+            render,          // Whether we whould render a partial or not
+            tidy,            // Should we tidy the ajax call output
+            credentials,     // Whether we should send cookie data with the call or not
+            headers          // What headers to use for the call
+        } = Object.assign({}, this.params, params );
 
-		if ( instance.upload ) {
-			options.xhr = function() {
-				var xhr = new window.XMLHttpRequest();
 
-				xhr.upload.addEventListener( 'progress', instance.uploadProgressHandler, false );
-				xhr.addEventListener( 'progress', instance.downloadProgressHandler, false );
-			};
-		}
+        // Edge doesnt implement the "finally" function in its fetch but does in a Promise so we need to wrap this in a Promise
+        return new Promise( ( resolve, reject ) => {
+            const fetchParams = {
+                method,
+                body: JSON.stringify({
+                    dustpress_data: {
+                        args,
+                        bypassMainQuery,
+                        partial,
+                        path,
+                        render,
+                        tidy,
+                        token
+                    }
+                }),
+                credentials,
+                headers
+            };
 
-		instance.start();
+            // Send the request
+            fetch( url, fetchParams ).then( ( response ) => {
+                if ( response.ok ) {
+                    return response.json();
+                }
 
-		instance.successHandler = function( data, textStatus, jqXHR ) {
-			var parsed;
+                // Reject promise as well if the response was not 200
+                throw Error( response.statusText );
+            }).then( ( data ) => {
+                resolve( data );
+            }).catch( ( err ) => {
+                reject( err );
+            });
+        });
+    }
+}
 
-			if ( typeof data === 'string' ) {
-				parsed = $.parseJSON( data );
-			}
-			else {
-				parsed = data;
-			}
-			
-			// Expire CSRF cookie
-			document.cookie = 'dpjs_token=; expires=-1; path=/';
-	
-			// Add to debugger data if it exists
-			if ( window.DustPressDebugger ) {
-				delete instance.params.success;
-				delete instance.params.error;
-	
-				var debug = {
-					params: instance.params,
-					data: parsed.debug ? parsed.debug : parsed
-				};
-				window.DustPressDebugger.extend(debug, instance.path);
-			}
-	
-			if ( parsed && ! parsed.error ) {
-				instance.success(parsed.success, parsed.data, textStatus, jqXHR);
-			}
-			else {
-				instance.error(parsed, textStatus, jqXHR);
-			}
-		};
-	
-		instance.errorHandler = function( jqXHR, textStatus, errorThrown ) {
-			// Expire CSRF cookie
-			document.cookie = 'dpjs_token=; expires=-1; path=/';
-	
-			instance.error({error: errorThrown}, textStatus, jqXHR);
-		};
-	
-		instance.uploadProgressHandler = function( event ) {
-			if ( event.lengthComputable ) {
-				var complete = ( event.loaded / event.total );
-	
-				instance.uploadProgress( complete );	
-			}
-		};
-	
-		instance.downloadProgressHandler = function( event ) {
-			if ( event.lengthComputable ) {
-				var complete = ( event.loaded / event.total );
-	
-				instance.downloadProgress( complete );
-			}
-		};
-
-		return $.ajax( options )
-		.done( instance.successHandler )
-		.fail( instance.errorHandler )
-		.always( instance.complete );
-	};
-
-	return dp;
-
-})( window, document, jQuery );
-
-window.dp = window.DustPress.ajax;
+// Setup global dp() function
+if ( typeof window.dp === 'undefined' ) {
+    const dustpress = new DustPress();
+    window.dp = ( path, params ) => dustpress.send( path, params );
+}