+++ webpack.config.js
@@ -1,11 +1,3 @@
-const webpack = require( 'webpack' );
-
-// Provide jQuery instance for all modules.
-const providePlugin = new webpack.ProvidePlugin({
-    jQuery: 'jquery',
-    '$': 'jquery'
-});
-
 /**
  * Webpack options
  * @type {Object}
@@ -30,17 +22,12 @@ module.exports = {
                 use: {
                     loader: 'babel-loader',
                     options: {
-                      presets: [ '@babel/preset-env' ]
+                        presets: [ '@babel/preset-env' ],
+                        plugins: [ '@babel/plugin-proposal-class-properties' ]
                     }
                 }
             }
         ]
-    },
-    plugins: [
-        providePlugin
-    ],
-    externals: {
-        'jquery': 'jQuery'
     },
     watchOptions: {
         poll: 100