+++ plugin.php
@@ -3,8 +3,8 @@
  * Plugin Name: DustPress.JS
  * Plugin URI: https://github.com/devgeniem/dustpress-js
  * Description: DustPress JavaScript library. Provides a front-end interface for interacting with the backend functions.
- * Version: 3.0.5
- * Author: Geniem Oy / Miika Arponen & Ville Siltala
+ * Version: 4.0.0
+ * Author: Geniem Oy / Miika Arponen, Ville Siltala & Arttu Mäkipörhölä
  * Author URI: http://www.geniem.com
  */
 class DustPressJs {