+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Changed the format of dustpress.js to a class
 - Use fetch instead of jquery
+- Use same CSRF token for all requests in the same session
 
 ## [3.0.4] - 2018-08-09